Combine and simplify some LGV components
These components used to have scrolling logic, but now that they do not,
they're not justified being a separate component.
