+++ packages/linear-genome-view/src/BasicTrack/components/BlockBasedTrack.js
@@ -1,22 +1,37 @@
 import { getConf } from '@gmod/jbrowse-core/configuration'
+import { makeStyles } from '@material-ui/core/styles'
 import { observer, PropTypes as MobxPropTypes } from 'mobx-react'
 import { getParent } from 'mobx-state-tree'
 import PropTypes from 'prop-types'
 import React from 'react'
-import Track from './Track'
 import TrackBlocks from './TrackBlocks'
 
+const useStyles = makeStyles({
+  track: {
+    position: 'relative',
+    whiteSpace: 'nowrap',
+    textAlign: 'left',
+    width: '100%',
+    minHeight: '100%',
+  },
+})
+
 function BlockBasedTrack(props) {
+  const classes = useStyles()
   const { model, children } = props
   return (
-    <Track {...props} trackId={getConf(model, 'trackId')}>
+    <div
+      data-testid={`track-${getConf(model, 'trackId')}`}
+      className={classes.track}
+      role="presentation"
+    >
       {model.trackMessageComponent ? (
         <model.trackMessageComponent model={model} />
       ) : (
         <TrackBlocks {...props} viewModel={getParent(getParent(model))} />
       )}
       {children}
-    </Track>
+    </div>
   )
 }
